Needed to recompile "blobslib" to Windows, so there was some adjusts to adequate the code. Changed the sin, at the lineFitter as there was some problems when compiling with VS 2008
git-svn-id: http://ccv-hand.googlecode.com/svn/trunk@80 f19414bc-3f48-11de-8f11-b9fadefa0f10

+++ CCV-HAND_Windows/addons/ofxNCore/src/Tracking/hands/videoHandFingerDetector.cpp
@@ -141,9 +141,13 @@ void videoHandFingerDetector::finger(handBlob	&blob) {
 		point3 = blob.myBlob.pts[i+2];
 
 		if(point2.y < point1.y && point2.y < point3.y) {
-			int lineAC1 = pow(((pow((point1.y - point2.y),2)) + (pow((point2.x - point1.x),2))),(1/2));
+			/**changed line A1 to float for VS 2008 compilation
+			*/
+			float lineAC1 = pow(((pow((point1.y - point2.y),2)) + (pow((point2.x - point1.x),2))),(1/2));
 			float angleA1 = acos(sin(lineAC1));
-			int lineAC2 = pow(((pow((point3.y - point2.y),2))+(pow((point3.x - point2.x),2))),(1/2));
+			/**changed line A2 to float for VS 2008 compilation
+			*/
+			float lineAC2 = pow(((pow((point3.y - point2.y),2))+(pow((point3.x - point2.x),2))),(1/2));
 			float angleA2 = acos(sin(lineAC2));
 			float angle = angleA1 + angleA2;
 			//printf("Hoek %f \n", hoek);
